# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Talbatiyk",
        "version": "1.0.0",
        "description": "# Talbatiyk API\n\n\u0648\u0627\u062c\u0647\u0629 \u0628\u0631\u0645\u062c\u0629 \u062a\u0637\u0628\u064a\u0642\u0627\u062a \u0645\u0646\u0635\u0629 **\u0637\u0644\u0628\u064a\u062a\u0643**.\n\n\u0627\u0644\u0625\u0635\u062f\u0627\u0631 \u0627\u0644\u062d\u0627\u0644\u064a:\n\n`v1`\n\n\u062a\u0634\u0645\u0644 \u062d\u0627\u0644\u064a\u064b\u0627:\n\n- Authentication API.\n- Business API.\n- Business Locations API.\n- Business Contacts API.\n- Sanctum Bearer Token authentication.\n\n\u0633\u064a\u062a\u0645 \u062a\u0648\u0633\u064a\u0639\u0647\u0627 \u0644\u0627\u062d\u0642\u064b\u0627 \u0644\u062a\u0634\u0645\u0644:\n\n- \u0627\u0644\u0642\u062f\u0631\u0627\u062a \u0627\u0644\u062a\u062c\u0627\u0631\u064a\u0629.\n- \u0623\u0648\u0642\u0627\u062a \u0627\u0644\u0639\u0645\u0644.\n- \u0627\u0644\u062a\u0648\u062b\u064a\u0642.\n- \u0627\u0644\u0645\u062a\u0627\u0628\u0639\u0627\u062a.\n- \u0627\u0644\u0645\u0646\u062a\u062c\u0627\u062a.\n- \u0627\u0644\u0637\u0644\u0628\u0627\u062a.\n- \u0627\u0644\u0645\u0632\u0627\u0645\u0646\u0629 \u0645\u0639 \u062a\u0637\u0628\u064a\u0642 Flutter."
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/businesses/{business}/contacts": {
            "get": {
                "operationId": "businessContact.indexBusiness",
                "summary": "\u0639\u0631\u0636 \u0648\u0633\u0627\u0626\u0644 \u0627\u0644\u0627\u062a\u0635\u0627\u0644 \u0627\u0644\u0639\u0627\u0645\u0629 \u0644\u0644\u0646\u0634\u0627\u0637",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BusinessContactResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "post": {
                "operationId": "businessContact.storeBusiness",
                "summary": "\u0625\u0646\u0634\u0627\u0621 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u0639\u0627\u0645\u0629",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBusinessContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            }
        },
        "/businesses/{business}/contacts/{contact}": {
            "get": {
                "operationId": "businessContact.showBusiness",
                "summary": "\u0639\u0631\u0636 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u0639\u0627\u0645\u0629 \u0648\u0627\u062d\u062f\u0629",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "businessContact.updateBusiness",
                "summary": "\u062a\u0639\u062f\u064a\u0644 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u0639\u0627\u0645\u0629",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBusinessContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "businessContact.destroyBusiness",
                "summary": "\u062d\u0630\u0641 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u0639\u0627\u0645\u0629",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/businesses/{business}/contacts/{contact}/primary": {
            "post": {
                "operationId": "businessContact.setPrimaryBusiness",
                "summary": "\u062a\u0639\u064a\u064a\u0646 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u0639\u0627\u0645\u0629 \u0643\u0648\u0633\u064a\u0644\u0629 \u0631\u0626\u064a\u0633\u064a\u0629 \u0645\u0646 \u0646\u0648\u0639\u0647\u0627",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/businesses/{business}/locations/{location}/contacts": {
            "get": {
                "operationId": "businessContact.indexLocation",
                "summary": "\u0639\u0631\u0636 \u0648\u0633\u0627\u0626\u0644 \u0627\u0644\u0627\u062a\u0635\u0627\u0644 \u0627\u0644\u062e\u0627\u0635\u0629 \u0628\u0641\u0631\u0639",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BusinessContactResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "post": {
                "operationId": "businessContact.storeLocation",
                "summary": "\u0625\u0646\u0634\u0627\u0621 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u062e\u0627\u0635\u0629 \u0628\u0641\u0631\u0639",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBusinessContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            }
        },
        "/businesses/{business}/locations/{location}/contacts/{contact}": {
            "get": {
                "operationId": "businessContact.showLocation",
                "summary": "\u0639\u0631\u0636 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u062e\u0627\u0635\u0629 \u0628\u0641\u0631\u0639",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "businessContact.updateLocation",
                "summary": "\u062a\u0639\u062f\u064a\u0644 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u062e\u0627\u0635\u0629 \u0628\u0641\u0631\u0639",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBusinessContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BusinessContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "businessContact.destroyLocation",
                "summary": "\u062d\u0630\u0641 \u0648\u0633\u064a\u0644\u0629 \u0627\u062a\u0635\u0627\u0644 \u062e\u0627\u0635\u0629 \u0628\u0641\u0631\u0639",
                "tags": [
                    "BusinessContact"
                ],
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "BusinessContactResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "label": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_primary": {
                        "type": "boolean"
                    },
                    "is_verified": {
                        "type": "boolean",
                        "description": "\u0644\u0627 \u0646\u0631\u0633\u0644 verified_at \u0641\u0642\u0637 \u0643\u0642\u064a\u0645\u0629 \u0645\u0646\u0637\u0642\u064a\u0629\u060c\n\u0628\u0644 \u0646\u0631\u0633\u0644 \u0627\u0644\u062d\u0627\u0644\u0629 \u0648\u0627\u0644\u062a\u0627\u0631\u064a\u062e \u0644\u0644\u0627\u0633\u062a\u0641\u0627\u062f\u0629 \u0645\u0646\u0647\u0645\u0627 \u0645\u0633\u062a\u0642\u0628\u0644\u064b\u0627."
                    },
                    "verified_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "type",
                    "value",
                    "label",
                    "is_primary",
                    "is_verified",
                    "verified_at"
                ],
                "title": "BusinessContactResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}